fix(ios): make InputParameterData accept value types besides string
This requires the code to validate the type being analysed.
Considering that we want to split between the available types (Decimal and String), we declare the 'OSFANLInputDataFieldKey' values that should have a Decimal value and be treated differently.
These changes allow us to
- refactor the code a bit, namely the 'OSFANLManager''s 'validate(dataField:isRequired:eventData:)' method, making it much simpler than it was.
- remove the now unused 'StringConvertable' initialiser.

References: https://outsystemsrd.atlassian.net/browse/RMET-3823