fix: always set 'FirebaseAutomaticScreenReportingEnabled' to 'true'
Context: The current version of the plugin for Cordova always sets this to true using the plugin.xml. We should keep the same behaviour, so we should set it to true. We need to set it through the YAML and not plugin.xml because for iOS it doesn't work setting it through the plugin.xml for Capacitor apps.

References: https://outsystemsrd.atlassian.net/browse/RMET-3643