feat: add 'google_analytics_automatic_screen_reporting_enabled' in YAML as well
Context: We already add it to the AndroidManifest.xml file of 'capacitor-cordova-android-plugins' using a config-file entry in the plugin.xml, but we might as well add it to the app's AndroidManifest.xml file by setting it in the build action. There won't be a problem of writing the same value twice because it always has the same value (true).

References: https://outsystemsrd.atlassian.net/browse/RMET-3643